perf(netlify,netlify-edge): exclude static paths from SSR function
The netlify preset was already using `preferStatic: true` which lets existing static
assets take precedence over the function, but this now also avoids a function invocation
for static paths that don't exist, i.e. this avoids an unnecessary dynamic 404 that could
be served directly from the CDN.

The `netlify-edge` preset wasn't excluding anything, so this addresses both the 404 case
and the existing asset case.

The 404 case is important because browsers frequently attempt to request hashed assets
from previous deploys that have been invalidated. There's no reason for this to go through
functions, since we know that the whole parent path is static.